move Job and Component classes to separate python files
